# html/sibapi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manejo de información mediante el uso de API",
        "contact": {
            "email": "admin@seth.com.bo"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/esquila/": {
            "get": {
                "tags": [
                    "VICUNA"
                ],
                "summary": "Lista de actas de esquila de Vicuña",
                "description": "Devuelve la lista de los datos de actas de equila de vicuña",
                "operationId": "v1getEsquilaList",
                "responses": {
                    "200": {
                        "description": "Json con datos de la lista de actas de equila de vicuña",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autentificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/esquila/{esquilaId}": {
            "get": {
                "tags": [
                    "VICUNA"
                ],
                "summary": "Buscar una acta de esquila de vicuña por ID",
                "description": "Devuelve la información de una acta de esquila de vicuña específico",
                "operationId": "v1getEsquila",
                "parameters": [
                    {
                        "name": "citesId",
                        "in": "path",
                        "description": "ID de la solicitud acta de esquila de vicuña",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Se optiene los datos con éxito",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron datos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autentificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/icas_institucion/": {
            "get": {
                "tags": [
                    "ICAS"
                ],
                "summary": "Lista de instituciones ICAS",
                "description": "Devuelve la lista de los datos de instituciones ICAS",
                "operationId": "v1getIcasInstitucionList",
                "responses": {
                    "200": {
                        "description": "Json con datos de la lista de instituciones ICAS",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autentificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/icas_institucion/{icasInsitucionId}": {
            "get": {
                "tags": [
                    "ICAS"
                ],
                "summary": "Buscar una institución por ID",
                "description": "Devuelve la información de una institución (ICAS) específico",
                "operationId": "v1getIcasInstitucion",
                "parameters": [
                    {
                        "name": "citesId",
                        "in": "path",
                        "description": "ID de la institución",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Se optiene los datos con éxito",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron datos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autentificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/lagarto_institucion/": {
            "get": {
                "tags": [
                    "LAGARTO"
                ],
                "summary": "Lista de instituciones de lagarto",
                "description": "Devuelve la lista de los datos de instituciones de lagarto",
                "operationId": "v1getLagartoInstitucionList",
                "responses": {
                    "200": {
                        "description": "Json con datos de la lista de institcuiones de lagarto",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autentificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/lagarto_institucion/{lagartoInsitucionId}": {
            "get": {
                "tags": [
                    "LAGARTO"
                ],
                "summary": "Buscar una institución por ID",
                "description": "Devuelve la información de una institución (LAGARTO) específico",
                "operationId": "v1getLagartoInstitucion",
                "parameters": [
                    {
                        "name": "citesId",
                        "in": "path",
                        "description": "ID de la institución",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Se optiene los datos con éxito",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron datos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autentificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authorization token optienes en 'login' - Auth: /api/login ",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}